+++ src/RcppExports.cpp
@@ -643,3 +643,16 @@ BEGIN_RCPP
     return rcpp_result_gen;
 END_RCPP
 }
+// rotatePointMatrix_
+NumericMatrix rotatePointMatrix_(NumericMatrix coordMat, NumericVector angleVec, NumericVector originVec);
+RcppExport SEXP _icosa_rotatePointMatrix_(SEXP coordMatSEXP, SEXP angleVecSEXP, SEXP originVecSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< NumericMatrix >::type coordMat(coordMatSEXP);
+    Rcpp::traits::input_parameter< NumericVector >::type angleVec(angleVecSEXP);
+    Rcpp::traits::input_parameter< NumericVector >::type originVec(originVecSEXP);
+    rcpp_result_gen = Rcpp::wrap(rotatePointMatrix_(coordMat, angleVec, originVec));
+    return rcpp_result_gen;
+END_RCPP
+}