+++ tests/main.R
@@ -18,5 +18,9 @@ parallel::clusterCall(cl, source, "icosa/tests/source.R")
 # occupied paleomap
 paleomap <- run_test_dir("icosa/tests/paleomap")
 
+
+# spherical tests
+point_rot <- run_test_file("icosa/tests/_spherical/point_rotation.R")
+
 # Finish
 stopCluster(cl)